refactor(isc-frontend): create form component and styles hook
