+++ isc-frontend/src/components/form-page/index.js
@@ -1,0 +1,1 @@
+export { FormPage } from "./form-page";
