+++ isc-frontend/src/components/index.js
@@ -3,3 +3,4 @@ export { Sidebar } from "./side-bar";
 export { Auth } from "./authentication";
 export { Header } from "./header";
 export { ContentWrapper } from "./content-wrapper";
+export { FormPage } from "./form-page";
